+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 - Search users by fullname and email. @alecghica
 
+- Fix login on Volto frontend when already logged-in in Plone Classic. @avoinea
+
 
 ## 1.2.0 (2023-09-13)