+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.2.1 (unreleased)
 
+- Search users by fullname and email. @alecghica
+- Fix login on Volto frontend when already logged-in in Plone Classic. @avoinea
 - Add the possibility to override the ZopeRequestAdapter.