+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Search users by fullname and email. @alecghica
 - Fix login on Volto frontend when already logged-in in Plone Classic. @avoinea
 - Add the possibility to override the ZopeRequestAdapter.
+- Fix the authomatic view when it is reporting an exception that does not have a message attribute
 
 
 ## 1.2.0 (2023-09-13)